More extensible multiaddress format
