+++ frame/indices/src/lib.rs
@@ -22,6 +22,7 @@
 
 use sp_std::prelude::*;
 use codec::Codec;
+use sp_runtime::MultiAddress;
 use sp_runtime::traits::{
 	StaticLookup, Member, LookupError, Zero, Saturating, AtLeast32Bit
 };
@@ -30,15 +31,12 @@ use frame_support::dispatch::DispatchResult;
 use frame_support::traits::{Currency, ReservableCurrency, Get, BalanceStatus::Reserved};
 use frame_support::weights::Weight;
 use frame_system::{ensure_signed, ensure_root};
-use self::address::Address as RawAddress;
 
 mod mock;
-pub mod address;
 mod tests;
 mod benchmarking;
 mod default_weights;
 
-pub type Address<T> = RawAddress<<T as frame_system::Trait>::AccountId, <T as Trait>::AccountIndex>;
 type BalanceOf<T> = <<T as Trait>::Currency as Currency<<T as frame_system::Trait>::AccountId>>::Balance;
 
 pub trait WeightInfo {
@@ -295,17 +293,18 @@ impl<T: Trait> Module<T> {
 
 	/// Lookup an address to get an Id, if there's one there.
 	pub fn lookup_address(
-		a: address::Address<T::AccountId, T::AccountIndex>
+		a: MultiAddress<T::AccountId, T::AccountIndex>
 	) -> Option<T::AccountId> {
 		match a {
-			address::Address::Id(i) => Some(i),
-			address::Address::Index(i) => Self::lookup_index(i),
+			MultiAddress::Id(i) => Some(i),
+			MultiAddress::Index(i) => Self::lookup_index(i),
+			_ => None,
 		}
 	}
 }
 
 impl<T: Trait> StaticLookup for Module<T> {
-	type Source = address::Address<T::AccountId, T::AccountIndex>;
+	type Source = MultiAddress<T::AccountId, T::AccountIndex>;
 	type Target = T::AccountId;
 
 	fn lookup(a: Self::Source) -> Result<Self::Target, LookupError> {
@@ -313,6 +312,6 @@ impl<T: Trait> StaticLookup for Module<T> {
 	}
 
 	fn unlookup(a: Self::Target) -> Self::Source {
-		address::Address::Id(a)
+		MultiAddress::Id(a)
 	}
 }
